+++ src/onegov.town/onegov/town/views/allocation.py
@@ -1,7 +1,9 @@
 import morepath
 
+from itertools import groupby
 from libres.db.models import Allocation
 from libres.modules.errors import LibresError
+from operator import attrgetter
 from onegov.core.security import Public, Private
 from onegov.libres import Resource, ResourceCollection
 from onegov.town import TownApp, _, utils
@@ -30,13 +32,32 @@ def view_allocations_json(self, request):
         return []
 
     scheduler = self.get_scheduler(request.app.libres_context)
+    queries = scheduler.queries
 
-    query = scheduler.allocations_in_range(start, end)
+    # get all allocations (including mirrors), for the availability calculation
+    query = scheduler.allocations_in_range(start, end, masters_only=False)
     query = query.order_by(Allocation._start)
 
-    return [
-        utils.AllocationEventInfo(a, request).as_dict() for a in query.all()
-    ]
+    allocations = query.all()
+
+    # put only return the master allocations
+    events = []
+
+    for key, group in groupby(allocations, key=attrgetter('_start')):
+        grouped = tuple(group)
+        availability = queries.availability_by_allocations(grouped)
+
+        for allocation in grouped:
+            if allocation.is_master:
+                events.append(
+                    utils.AllocationEventInfo(
+                        allocation,
+                        availability,
+                        request
+                    )
+                )
+
+    return [e.as_dict() for e in events]
 
 
 def get_new_allocation_form_class(resource, request):